Clan system, db fixes

+++ cogs/privaterooms.py
@@ -1,5 +1,6 @@
 import logging
 import discord
+import datetime
 import settings
 
 from discord import app_commands, VoiceState, Member
@@ -8,7 +9,8 @@ from discord.ext import commands
 from views.privaterooms import PrivateRoomsView
 
 from sqlalchemy import select
-from data.db_session import create_session, PrivateRoom
+from data.db_session import create_session
+from data.tables import Player, PrivateRoom
 
 _log = logging.getLogger(__name__)
 
@@ -35,38 +37,64 @@ class PrivateRoomsCog(commands.Cog):
 
     @commands.Cog.listener()
     async def on_voice_state_update(self, member: Member, before: VoiceState, after: VoiceState):
-        if after.deaf or after.mute or after.self_deaf or after.self_mute \
-                or before.deaf or before.mute or before.self_deaf or before.self_mute:
+        # Put context manager here?
+        if not before.channel and after.channel:
+            async with create_session() as session:
+                async with session.begin():
+                    result = await session.execute(select(Player).where(Player.id == member.id))
+                    player: Player = result.scalars().one_or_none()
+                    player.joined_vc = datetime.datetime.now().replace().replace(microsecond=0)
+
+        if before.channel and not after.channel:
+            async with create_session() as session:
+                async with session.begin():
+                    result = await session.execute(select(Player).where(Player.id == member.id))
+                    player: Player = result.scalars().one_or_none()
+                    player.last_seen = datetime.datetime.now().replace(microsecond=0)
+
+                    # if member was in vc when bot was reloaded
+                    if player.joined_vc:
+                        player.voice_activity += player.last_seen - player.joined_vc
+
+        # Replace with any()?
+        # Checks if member used mute, deaf etc.
+        if after.deaf or after.mute or \
+                after.self_deaf or after.self_mute or \
+                after.self_video or after.self_stream or \
+                before.deaf or before.mute or \
+                before.self_deaf or before.self_mute or \
+                before.self_video or before.self_stream:
             return
 
-        # Delete voice channel after owner left
-        if before.channel:
-            if before.channel.id != 1153441991980494879 and before.channel.category.id == 1153441782953160795:
+        if before.channel and before.channel.category:
+            if before.channel.category.id == settings.PRIVATE_ROOMS_CATEGORY_ID and len(before.channel.members) == 0:
                 async with create_session() as session:
                     async with session.begin():
-                        room = await session.execute(
+                        result = await session.execute(
                             select(PrivateRoom)
-                            .where(PrivateRoom.owner == member.id)
-                            .where(PrivateRoom.id == before.channel.id))
-                        room = room.scalars().one_or_none()
-                        if room is not None:
+                            .where(PrivateRoom.id == before.channel.id)
+                        )
+                        room: PrivateRoom = result.scalars().one_or_none()
+                        if room:
                             await session.delete(room)
                             await before.channel.delete()
 
-        # Create channel when member join voice channel
-        if after.channel and after.channel.id == 1153441991980494879:
-            channel = await after.channel.category.create_voice_channel(name=member.name)
-
+        # Create channel when member join create voice channel
+        if after.channel and after.channel.id == settings.ENTRY_ROOM_ID:
             async with create_session() as session:
                 async with session.begin():
-                    session.add(PrivateRoom(channel.id, member.id))
+                    room = await PrivateRoom.get_room(member.id)
+                    if room:
+                        await member.edit(voice_channel=self._bot.get_channel(room.id))
+                        return
 
-            await member.move_to(channel)
-            return
+                    channel = await after.channel.category.create_voice_channel(name=f'{member.name}')
+                    session.add(PrivateRoom(channel.id, member.id))
+                    await member.move_to(channel)
 
     @commands.Cog.listener()
     async def on_message(self, message: discord.Message):
-        if message.channel.id == 1153441934128467968 and not message.author.bot:
+        if message.channel.id == settings.MANAGE_CHANNEL_ID and not message.author.bot:
             await message.delete()
